+++ core/src/jsMain/kotlin/dev/fritz2/dom/comment.kt
@@ -1,20 +1,39 @@
 package dev.fritz2.dom
 
+import dev.fritz2.dom.html.RenderContext
 import kotlinx.browser.document
+import kotlinx.coroutines.cancelChildren
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.map
 import org.w3c.dom.Node
 
 /**
  * Provides functionality to handle comments.
  */
-interface WithComment<out T : Node> : WithDomNode<T> {
+interface WithComment<out T : Node> : WithDomNode<T>, RenderContext {
 
     /**
-     * Adds a comment at this position.
+     * Adds comment-content of a [Flow] at this position
      *
      * @receiver comment-content
      */
-    fun String.asComment() {
-        domNode.appendChild(document.createComment(this))
+    fun Flow<String>.asComment() {
+        mountDomNode(job, domNode) { childJob ->
+            childJob.cancelChildren()
+            this.map { TextNode(it) }
+        }
+    }
+
+    /**
+     * Adds comment-content of a [Flow] at this position
+     *
+     * @receiver comment-content
+     */
+    fun <T> Flow<T>.asComment() {
+        mountDomNode(job, domNode) { childJob ->
+            childJob.cancelChildren()
+            this.map { TextNode(it.toString()) }
+        }
     }
 
     /**
