+++ core/src/jsTest/kotlin/dev/fritz2/binding/mount.kt
@@ -3,7 +3,6 @@ package dev.fritz2.binding
 import dev.fritz2.dom.html.render
 import dev.fritz2.dom.mount
 import dev.fritz2.identification.uniqueId
-import dev.fritz2.test.checkFlow
 import dev.fritz2.test.checkSingleFlow
 import dev.fritz2.test.initDocument
 import dev.fritz2.test.runTest
@@ -66,10 +65,8 @@ class MountTests {
         render {
             div(id = outer) {
                 text.render {
-                    render {
-                        div(id = inner1) {
-                            +it
-                        }
+                    div(id = inner1) {
+                        +it
                     }
                 }
                 div(id = inner2) {
