+++ core/src/jsTest/kotlin/dev/fritz2/dom/comment.kt
@@ -19,15 +19,11 @@ class CommentTests {
         initDocument()
 
         val id1 = uniqueId()
-        val id2 = uniqueId()
         val comment = "testComment"
 
         render {
             section {
                 div(id = id1) {
-                    comment.asComment()
-                }
-                div(id = id2) {
                     !comment
                 }
             }
@@ -40,11 +36,5 @@ class CommentTests {
         assertEquals(id1, div1.id)
         assertEquals(8, div1.firstChild?.nodeType)
         assertEquals(comment, div1.firstChild?.nodeValue)
-
-        val div2 = document.getElementById(id2) as HTMLDivElement
-
-        assertEquals(id2, div2.id)
-        assertEquals(8, div2.firstChild?.nodeType)
-        assertEquals(comment, div2.firstChild?.nodeValue)
     }
 }
